pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clinica-assistant"
version = "0.1.0"
description = "Webhook clínico para Chatwoot con LangGraph, vLLM y memoria."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "httpx>=0.28.0,<1.0.0",
  "langgraph>=0.3.0,<0.4.0",
  "mem0ai>=0.1.82",
  "pydantic>=2.10.0,<3.0.0",
  "pydantic-settings>=2.7.0,<3.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.25.0,<1.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
